Processing RIC Indication trigger

+++ e2sim/e2sm_examples/kpm_e2sm/src/ASN1c/RICrequestID.h
@@ -21,7 +21,7 @@ extern "C" {
 
 /* RICrequestID */
 typedef struct RICrequestID {
-	long	 ricRequestorID;
+	int	 ricRequestorID;
 	long	 ricInstanceID;
 	/*
 	 * This type is extensible,
